+++ interface-impl-amf/src/main/java/org/mule/amf/impl/validation/JsonParameterValidationStrategy.java
@@ -4,8 +4,10 @@
  * license, a copy of which has been included with this distribution in the
  * LICENSE.txt file.
  */
-package org.mule.amf.impl.model;
+package org.mule.amf.impl.validation;
 
+import amf.client.model.domain.AnyShape;
+import amf.client.model.domain.Shape;
 import amf.client.model.domain.ValidatorAware;
 import amf.client.validate.PayloadValidator;
 import amf.client.validate.ValidationReport;
@@ -13,8 +15,10 @@ import org.json.simple.JSONValue;
 import org.mule.amf.impl.exceptions.ParserException;
 import org.mule.amf.impl.util.LazyValue;
 
-import static org.mule.amf.impl.model.MediaType.APPLICATION_JSON;
-import static org.mule.amf.impl.model.MediaType.APPLICATION_YAML;
+import static org.mule.amf.impl.util.AMFUtils.APPLICATION_JSON;
+import static org.mule.amf.impl.util.AMFUtils.APPLICATION_YAML;
+import static org.mule.amf.impl.util.AMFUtils.needsQuotes;
+
 
 class JsonParameterValidationStrategy extends ValidationStrategy {
 
@@ -31,6 +35,10 @@ class JsonParameterValidationStrategy extends ValidationStrategy {
 
   public JsonParameterValidationStrategy(ValidatorAware validatorAware, boolean schemaNeedsQuotes) {
     super(validatorAware, schemaNeedsQuotes);
+  }
+
+  public JsonParameterValidationStrategy(AnyShape shape) {
+    super(shape, needsQuotes(shape));
   }
 
   @Override
